chapitre 4 terminé

+++ backend/controllers/stuff.js
@@ -1,4 +1,6 @@
+const fs = require("fs");
 const Thing = require("../models/Thing");
+const { error } = require("console");
 
 exports.createThing = (req, res, next) => {
   const thingObject = JSON.parse(req.body.thing);
@@ -70,17 +72,20 @@ exports.modifyThing = (req, res, next) => {
 };
 
 exports.deleteThing = (req, res, next) => {
-  Thing.deleteOne({ _id: req.params.id })
-    .then(() => {
-      res.status(200).json({
-        message: "Deleted!",
-      });
+  Thing.findOne({ _id: req.params.id })
+    .then((thing) => {
+      if (thing.userId !== req.auth.userId) {
+        res.status(401).json({ message: "Action non autorisée ❗" });
+      } else {
+        const filename = thing.imageUrl.split("/images/")[1];
+        fs.unlink(`images/${filename}`, () => {
+          Thing.deleteOne({ _id: req.params.id })
+            .then(() => res.status(200).json({ message: "Objet supprimé" }))
+            .catch(() => res.status(401).json({ error }));
+        });
+      }
     })
-    .catch((error) => {
-      res.status(400).json({
-        error: error,
-      });
-    });
+    .catch(() => res.status(401).json({ error }));
 };
 
 exports.getAllStuff = (req, res, next) => {
